requirements.txt
# Requirements.txt

# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
PyYAML>=6.0.0

# ML
scikit-learn>=1.3.0
torch>=2.0.0
torchvision
torchaudio

# Visualization (옵션)
matplotlib>=3.7.0
seaborn>=0.12.2

# Dev Utils
tqdm>=4.65.0
ipykernel
jupyter
